requirements.txt
# Core ML/DL
torch>=2.0.0
transformers>=4.35.0
peft>=0.7.0
datasets>=2.15.0

# Classical ML
scikit-learn>=1.3.0
joblib>=1.3.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Web App
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
